% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots.R
\name{boxplots}
\alias{boxplots}
\title{Side-By-Side Box plots Function}
\usage{
boxplots(data, cat, num)
}
\arguments{
\item{data}{The tibble or data frame which contains the numeric
and text/factor columns.}

\item{cat}{The text or factor column name.}

\item{num}{The numeric column name.}
}
\value{
Box plots of the numeric column for each level of the
text/factor column. Throws an error if \code{cat} and \code{num} are not the proper class.
}
\description{
This function produces box plots for a numeric column
across levels of a text/factor column, from a tibble or data frame.
}
\examples{
boxplots(datateachr::apt_buildings, air_conditioning, year_built)
boxplots(gapminder::gapminder, continent, lifeExp)

}
